+++ packages/config/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @vegastack/pages-config
 
+## 0.1.7
+
 ## 0.1.6
 
 ## 0.1.4