+++ packages/db/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @vegastack/pages-db
 
+## 0.1.7
+
 ## 0.1.6
 
 ### Patch Changes